requirements.txt
asgiref==3.3.1
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.7
django-jalali-date==0.3.1
ipython==7.21.0
ipython-genutils==0.2.0
jdatetime==3.6.2
jedi==0.18.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.16
Pygments==2.8.1
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
